Move other shared objects to common
Moved to commons:
YoutubeDlRequest
YoutubDlOption
YoutubeDlResponse
