+++ commons/src/main/java/com/farimarwat/commons/YoutubeDLOptions.kt
@@ -1,4 +1,4 @@
-package com.farimarwat.library
+package com.farimarwat.commons
 
 import kotlin.collections.ArrayList
 import kotlin.collections.LinkedHashMap
