+++ commons/src/main/java/com/farimarwat/commons/YoutubeDLRequest.kt
@@ -1,4 +1,4 @@
-package com.farimarwat.library
+package com.farimarwat.commons
 
 class YoutubeDLRequest {
     private val urls: List<String>
